+++ pennyway-client-iOS/pennyway-client-iOS/Clean/Domain/Interfaces/LoginRepository.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 protocol LoginRepository {
-    func login(username: String, password: String, completion: @escaping (Result<AuthResponseDto, Error>) -> Void)
-    func oauthLogin(model: OAuthLogin, completion: @escaping (Result<AuthResponseDto, Error>) -> Void)
-    func checkLoginState(completion: @escaping (Result<AuthResponseDto, Error>) -> Void)
+    func login(username: String, password: String, completion: @escaping (Result<AuthResponseData, Error>) -> Void)
+    func oauthLogin(model: OAuthLogin, completion: @escaping (Result<AuthResponseData, Error>) -> Void)
+    func checkLoginState(completion: @escaping (Result<AuthResponseData, Error>) -> Void)
 }
