+++ pennyway-client-iOS/pennyway-client-iOS/Clean/Domain/Usecases/SignUpUseCase.swift
@@ -18,16 +18,18 @@ protocol SignUpUseCase {
 
 class DefaultSignUpUseCase: SignUpUseCase {
     private let repository: SignUpRepository
+    private let chatStompService = DefaultChatStompService.shared
 
     init(repository: SignUpRepository) {
         self.repository = repository
     }
 
     func signUp(model: SignUp, completion: @escaping (Bool, UserId?) -> Void) {
-        repository.signUp(model: model) { result in
+        repository.signUp(model: model) { [weak self] result in
             switch result {
             case let .success(response):
-                completion(true, UserId(id: Int64(response.data.user.id)))
+                self?.chatStompService.connect()
+                completion(true, UserId(id: Int64(response.user.id)))
             case let .failure(error):
                 Log.error("Sign up failed: \(error)")
                 completion(false, nil)
@@ -36,11 +38,12 @@ class DefaultSignUpUseCase: SignUpUseCase {
     }
 
     func oauthSignUp(model: OAuthSignUp, completion: @escaping (Bool, UserId?) -> Void) {
-        repository.oauthSignUp(model: model) { result in
+        repository.oauthSignUp(model: model) { [weak self] result in
             switch result {
             case let .success(response):
+                self?.chatStompService.connect()
                 KeychainHelper.deleteOAuthUserData()
-                completion(true, UserId(id: Int64(response.data.user.id)))
+                completion(true, UserId(id: Int64(response.user.id)))
             case let .failure(error):
                 Log.error("OAuth sign up failed: \(error)")
                 completion(false, nil)
